+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `CurveLoopUtils` to create `CurveLoop` from `XYZ` points and `Curve`.
 ### Generator
 - Add `SelectElements<T>` and `SelectElementTypes<T>` in the generator to generate `Select` extension methods for `Element` and `ElementType`.
+- Update `Generator` with summary comments for `Select` extension methods.
 
 ## [0.4.0] / 2024-09-23 - 2025-03-22
 ### Features